Add "Extension Version" field to help widget
Assisting users with extensions is difficult because it is unclear
whether problems occur because of bugs in the extension or due to a
mismatch in versions when extensions work together. This commit adds
an "Extension Version" header on the helper widget that displays the
version number specified in the @DesignerComponent annotation if the
version number is greater than 0.

Closes #894

Change-Id: If7fdd49795e81f8ff784036ce5745de499aafbd9