Fix extension upgrade issues
An interaction between an upgrading an extension with a new version
number and a change to YaVersion results in being unable to open a
project with extensions. This fix does two things to address this:

1. Force a save of the project so that the version numbers of the
extension are persisted back to the server.
2. Check whether a component to be upgraded is an external component,
and if so, skip the upgrade step since we don't have an upgrader for
it anyway.

Fixes #996

Change-Id: I6d39f70b3cac80961ed538f5a2bd99ab81729a5e