The CloudDB Component
CloudDB is similar to the FirebaseDB component but does not require a
proprietary back-end database. Instead the back-end is based on the open
source Redis system which is a well used and scale-able data storage
solution.

With CloudDB you can store arbitrary data, including binary image data,
based on a “tag” which is a string. The “tag” is combined with
ProjectID (which defaults to the name of the current project, but can be
changed) so that multiple projects can use the same tags without
interfering with each other.

The domain name system name of the Redis server should be configured the
“RedisServer” property. This defaults to “DEFAULT” which is a server
provided default (more on that below). The “Token” property should be
the password required for the Redis Server. An “out of the box” Redis
server does not implement SSL, so the UseSSL property should be
unchecked.

In the DEFAULT setup, the server name is set to DEFAULT, UseSSL is
checked and the Token is automatically generated when you load a
project. In the DEFAULT case, a special modified Redis server is used
which implements multi-user features which provide isolation between
different MIT App Inventor programmers. It also uses SSL to protect
on-line communications from interception.

Author: Natalie Lao <natalie@csail.mit.edu>
Author: Jeffrey I. Schiller <jis@mit.edu>
Author: Joy Mitra <joymitro1989@gmail.com>

Change-Id: Ic7d7aa0f7337a5452d9c138920813deb797fac20