Fix issues with marker size breaking project load
Projects using relatively sized markers are unable to load correctly
because the width/height is set before the MockMap is ready. This
change introduces a needsUpdate flag that is set when this occurs. The
size is then updated when addToMap is called once the map has been
set. This also fixes a minor bug where markers with image assets do
not end up being the correct size in the designer after toggling
visibility.

Fixes #1136 and fixes #1137

Change-Id: Ic800b19241b9bcf33cd07db17308b7c61d13b81e