Provide for a second buildserver.
Provide for a second buildserver. If provided then the user is presented
a choice of two different build servers. This feature was introduced to
support building against two different versions of the Android SDK. As
of 8/1/2018 MIT is using this feature with the first buildserver
building Apps with a minSdk of 7 (but no targetSdk defined) to support
older devices. The second buildserver is configured with a minSdk of 14
and a targetSdk of 26 (along with the necessary component code
changes). This is required for new Apps submitted to the Google Play
Store. Having two buildservers provides us the flexibility to support
older devices while also ensuring that people can submit MIT App
Inventor created Apps to the Google Play Store.

Change-Id: Iae58f9ff4d112f8bc5d3203675f37cec7bf76920