Chunk large Messages to the Companion
WebRTC can only safely send messages of 16K. So we will only send up to
8K characters (ok, a margin of safety). Most projects do not need to
send YAIL forms larger then this, but a few do.

To “chunk” the code into smaller pieces would normally require changes
both on the browser side and the Companion side. However we wish to
avoid changing the Companion at this time. So instead, we break the
large YAIL forms into smaller YAIL forms that can each be evaled on the
device correctly.

Change-Id: I865e39cad3c55ea35b2a56582d50136f8303e90e