Allow for Any Marker, LineString, and Polygon with Maps
When one dynamically loads map content from the web, one might want to
manipulate that content. To do that today, you must add one of each of
the relevant types to a Map and make them invisible so that the "Any
..." tree items are shown. This change makes it so that as long as the
Map exists it will add the any component options for Marker (Point),
LineString, and Polygon components.

Change-Id: Ifb6837be89231c7f713a140318858da366113c21