De-Bounce the “Connect with Code” Button
De-Bounce the “Connect with Code” Button in the MIT AI2 Companion

The Companion's design is to setup communications with the user's
browser and get to work. Once this process starts, enough things are in
motion that it is best to quit the Companion and start a fresh copy if a
different code is needed.

If the same code is entered more then once, we just ignore the second
attempt. This often happens when someone scans a QR Code and then
presses the "Connect" Button because they do not know that they don't
have to do that. Effectively we are "de-bouncing"

Change-Id: I06b20c5d9cdb07999a380d9f877edd111e0357c5