Emit disable event code only if no active handler exists
The change that produced calls to EventDispatcher's unregister method
for disabled blocks did not account for the scenario where both an
active and inactive event handler exist for the same component-event
combination. This change makes it so that unregister calls are only
output if there isn't a matching active event handler.

Change-Id: If33b577f5af03c64f7abb8fe5b587d576886889b