Hide deprecated blocks
In earlier versions of App Inventor, the deprecated flag was stored as
the string 'true' or 'false'. More recently, we made the flag output a
Boolean rather than a string (i.e., `true` or `false`). However, the
code to handle creating the blocks for the drawer wasn't updated to
handle this scenario. Luckily, it was a sensible fallback such that
the block showed, but that means that deprecated blocks show when they
shouldn't. This change fixes this by checking for both the string
version (to support extensions that use the old format) and the
Boolean version.

Change-Id: I6ae272d8e36887ea4259a7948a56badb62f817d6