Treat 'global' keyword as untranslatable internally
In the block name we store globals as 'global foo' where foo is the
variable name. In Korean and Polish, the 'global' keyword has been
translated for the blocks, but interally we still call the variable
'global foo' even though the keyword shown to the user is
translated. However, because the functionality to unprefix names
references the translated name it causes bad code generation for
languages where the global keyword isn't the string 'global'. This
change introduces a GLOBAL_KEYWORD constant that is checked to
identify globals where the user locale may provide a translation.

Change-Id: Id18331206c28592c145bc2a1d915bcecc0799f13