Implement copy and paste support in the designer (#2151)
* Implement copy and paste support in the designer

This change adds support for copying and pasting components in the
designer. It handles the entire tree from a given component (and
supports multiple roots for if we allow multiple selection in the
future). Name substitution will occur if a component exists with the
same name when pasting. Top level blocks referencing the copied
components are also included in the operation. Holding the shift key
while pasting will disable pasting the event handlers.

Resolves #726
Resolves #840

Change-Id: Ica9637f8d07e7287cd698248c60533a5e10af752

* Allow "copying" a MockForm by copying all of its children

Change-Id: Ifbb3ade0cebf4a2285f79459ddd80a3359b02819

* Implement copy-and-paste of Form

Change-Id: I5f0df1b775b9b6ac4fb9bddf188ce5b7adae53e9

* Fix pasting of visible and nonvisible components

Change-Id: I034a853118b079551d80b123ecba2ed6708e5984

* Add documentation for copy and paste functionality

Change-Id: I7dd70e7ae66e8bd186f3bfb1059a68fffa931214