Add osmdroid.jar as dependency of Navigation component
The Navigation component makes use of org.osmdroid.util.GeoPoint, but
if it is included in a project without also including a Map component,
compiled apps fail due to the fact that GeoPoint is not in the final
binary. This change asserts that the Navigation component is dependent
on osmdroid.jar when building apps.

Change-Id: I1111c710e3b53a6dc85d65668ef140d6286ec585