Implement LegacyMode for File component
With the nb184 change, apps that previously worked on Android 10 may
no longer work if the app in question stored files based on the
external storage root (e.g., /sdcard). While Google isn't enforcing
read/write access to directories outside of the app-private path just
yet (this will be true in Android 11), we switched over most file
operations to use the app-private path anyway starting with Android
Q. New apps will read/write files without issue, however apps that are
used to using the old path on Android 10 have no way to read files
they may have written.

This change adds a LegacyMode property to File, which is False by
default, that allows a developer to access files relative to the root
of the external storage on Android 10 and higher. I've included a note
that the use of this property may cause apps to fail starting with
Android 11.

Change-Id: I4ab0d26e54d2a74535b4a691cfd63826543d29a4