Tune the CloudDB Component
Internally we use the Java “Jedis” library to communicate with
CloudDB. However this library is tuned for high speed low latency
connections. In particular it provides for a 2 second timeout on
transactions. This is not always sufficient in a wide area network
situations that a lot of devices will be.

Most transactions work, but when storing larger objects this 2 second
timeout can be exceeded. This change updates the timeout to be 20
seconds instead of 2. It also works around a minor issue in the Jedis
library that was causing the authentication (login) transaction to be
duplicated. This was not a real problem, but it still should be fixed,
and is in this change.

Change-Id: I00056d5bcf6f79fc783da2b22ed68c17e946e0e5