Mitigate Chrome 89 change
Chrome 89 implements RFC8285 by default. It implements a general
mechanism for RTP header extensions. We do not use these....

Unfortunately the WebRTC library we are using in the Companion does not
understand the extension request in the WebRTC offer message and refuses
the offer completely, resulting in WebRTC not working from Chrome
89 (and presumably onward).

This change removes the extension from the offer prior to sending it to
the Companion, so WebRTC works again.

We believe this is harmless, because we do not use audio or video over
WebRTC and this extension has no impact on the data channel (which we do
use).

Change-Id: I68f4ebaa2099ca0b4bbc7b1abd0a565dbde1f4e6